Put a sample property data, and fill out the Agents full sample data and add data of the properties to the database For buy and rent pages 	Add Search button 	fetch properties from database and showcase it (use the python and postgres to make the database and fetch the data from postgres) 	When Showcasing the property display the short details of property. 	When View details button is pressed it needs to show the details of that property and also suggests other similar filters for the user needs. 	and add a live filter to fetching system, when the price or property description on the filter is edited the fetched property needs to change accordingly. This also applies to the MAP as well, when the user touches the red mark that indicates a property that property details needs to be showcased and on the side the recommended and similar properties needs to be suggested.
On the sell page incorporate the image attachment field where users can upload property images.
and also after the submit listing button is touched it needs to check if there are unfilled fields and if there are all filled field it needs to notify the admin and post it, mentioning that the you have add you new Listings.

Incorporate Twilio/SendGrid for email and password authentication as well.
Add login pages and incorporate vie Google, Facebook and other popular email service
Add Admin page to the navigation bar
Add Login Button on the Navigation bar and make the sign up page and add essential forms.
Add a Logout button if user is signed in and wants to Log out
Add Payment Details form on the pricing listed in the pricing page
	Incorporate Ethiopian popular payment getways like Chapa, Telebirr, Appolo and others
On the Sell Page Add the Fields found on the Buy page so that people that sell can fill it also on the page add the option if the user wants to sell using an agent or by their self, if using an agent the system needs to recommend them to an agent that is fit for them and has a high star rate, if not the system needs to check whether the client's subscription is above free module, if so it will proceed to posting if not it needs to suggest them to fill in to a subscription that is higher than free or else it would not support them.